+++ CHANGELOG.md
@@ -16,6 +16,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
 
 ### Added
 
+- `From<Bps>` for `serial::Config`
 - `From<Into<Hertz>>` for `i2c::Mode`
 - `exti_rtic` example
 - Support for OpenDrain pin configuration on SPI CLK and MOSI pins